[Update] Page Stickers > child pages sticker1/2/3
