+++ src/pages/stickers/pack1.tsx
@@ -1,0 +1,6 @@
+const Gato1Page = () => {
+    return <h1>Sticker Gato 1</h1>;
+  };
+  
+  export default Gato1Page;
+
