+++ src/styles/cat.scss
@@ -77,21 +77,36 @@
 
 
 .catinfo {
-    display: flex;
+    // display: flex;
     flex-wrap: wrap;
+    display: flex;
+    //flex-direction: column;
+    //flex-basis: calc(50% - 5px);
     gap: 10px;
     width: 100%;
     height: 80%;
 
+
     .catinfoitem {
-        display: flex;
-        flex-direction: column;
-        flex-basis: calc(50% - 5px);
-        width: 50%;
+        flex: 1;
+        //width: 50%;
         //height: 50%;
+        background-color: rgba(240, 248, 255, 0.1);
         padding: 20px;
         border-radius: 10px;
-        box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+        box-shadow: 0 0 10px rgba(84, 84, 84, 0.5);
+        transition: all 0.3s ease-in-out;
+        @media screen and (max-width: 768px) {
+            display: flex;
+            flex-direction: column;
+            width: 100%;
+            height: 100%;
+        }
+        &:hover {
+            transform: scaleY(1.1);
+            box-shadow: 0 0 5px 0 rgba(255, 255, 255, 0.17);
+        }
+        
     }
     .catinfoitem > div:last-child {
         flex-basis: calc(50% - 10px);
@@ -173,7 +188,18 @@
     &:hover {
         border-radius: 10px 10px 0 0;
     }
+  }
 
+  .cat{
+    width: 70px;
+    height: 50px;
+    margin: -10px;
+    padding: -10px;
+    border: red solid 1px;
+    transition: all 0.3s ease-in-out;
+    &:hover {
+        transform: scale(2);
+    }
   }
 
   .itemspace {
